overlay: Add QS gradient [1/3]
This commit adds a possibility to make QuckSettings buttons and brightness slider background filled with gradient(instead of just solid) with help of an overlay. The user-facing switch is added to theming module and a package itself is added Lineage overlays. Also SystemUI is modified to factor out shape of brightness slider in order to minimize hardcoding in overlay.
The values of dimens in overlay are hardcoded as SystemUI does not publish it. While they do not overlayed anywhere in device/OS tree it may become problem in future.

The resulting QS screenshot can be found here https://imgur.com/a/RFdraYg

The changes done to following repos:
* frameworks/base
* vendor/overlay
* packages/apps/ThemePicker

Change-Id: I9de7cce659379e7cfc61cdb9e009367cfccf37f9